Move satellite test in dedicated source file
