+++ CMakeLists.txt
@@ -132,6 +132,7 @@ set(PICOQUIC_TEST_LIBRARY_FILES
     picoquictest/picoquic_lb_test.c
     picoquictest/pn2pn64test.c
     picoquictest/sacktest.c
+    picoquictest/satellite_test.c
     picoquictest/skip_frame_test.c
     picoquictest/socket_test.c
     picoquictest/splay_test.c
