+++ picoquictest/picoquictest.vcxproj
@@ -163,6 +163,7 @@
     <ClCompile Include="picoquic_lb_test.c" />
     <ClCompile Include="pn2pn64test.c" />
     <ClCompile Include="sacktest.c" />
+    <ClCompile Include="satellite_test.c" />
     <ClCompile Include="skip_frame_test.c" />
     <ClCompile Include="socket_test.c" />
     <ClCompile Include="cplusplus.cpp" />
