+++ picoquictest/picoquictest.vcxproj.filters
@@ -96,6 +96,9 @@
     <ClCompile Include="picoquic_lb_test.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="satellite_test.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="picoquictest.h">
